refactor: move icons to separate modules; revert SVG files; rename modules
