+++ libs/shared-components/src/mini-overlay/assets/icons/Arrow.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const Arrow = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="14" height="8" viewBox="0 0 14 8" fill="none">
+    <path
+      d="M1.5 1.25L7 6.75L12.5 1.25"
+      strokeWidth="1.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+)
+
+export default Arrow
