+++ libs/shared-components/src/mini-overlay/assets/icons/Close.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const Close = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="9" height="10" viewBox="0 0 9 10" fill="none">
+    <path
+      d="M8 1.5L1 8.5M1 1.5L8 8.5"
+      stroke="#7A818B"
+      strokeWidth="1.16667"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+)
+
+export default Close
