+++ libs/shared-components/src/mini-overlay/assets/icons/MutationFallbackIcon.tsx
@@ -1,0 +1,25 @@
+import React from 'react'
+
+const MutationFallbackIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="44" height="44" viewBox="0 0 44 44" fill="none">
+    <rect x="0.5" y="0.5" width="43" height="43" rx="21.5" fill="#F8F9FF" />
+    <rect x="0.5" y="0.5" width="43" height="43" rx="21.5" stroke="#E2E2E5" />
+    <g clipPath="url(#clip0_352_10)">
+      <path
+        d="M32 26.1425C32 27.0292 31.796 27.8418 31.3886 28.5809C30.9807 29.32 30.4355 29.9079 29.7522 30.3446C29.0689 30.7819 28.3075 31 27.4685 31C26.7251 31 26.0356 30.8265 25.4005 30.4796C24.765 30.1326 24.2374 29.6639 23.8181 29.0724C23.4102 29.6639 22.8887 30.1326 22.2536 30.4796C21.6181 30.8265 20.9286 31 20.1857 31C19.3585 31 18.6004 30.7819 17.9109 30.3446C17.2215 29.9079 16.673 29.32 16.2656 28.5809C15.8577 27.8423 15.6542 27.0292 15.6542 26.1425V18.4708C15.6542 17.9387 16.0569 17.507 16.5533 17.507C17.0497 17.507 17.4524 17.9387 17.4524 18.4708V26.1425C17.4524 26.6822 17.575 27.1738 17.821 27.6171C18.0666 28.0605 18.3965 28.414 18.8101 28.6773C19.2236 28.9411 19.6822 29.0724 20.1857 29.0724C20.6892 29.0724 21.1264 28.9471 21.5344 28.6966C21.9418 28.446 22.2688 28.1087 22.5144 27.6846C22.76 27.2605 22.8949 26.7918 22.919 26.2775V18.4708C22.919 17.9387 23.3217 17.507 23.8181 17.507H23.8361C24.3325 17.507 24.7352 17.9387 24.7352 18.4708V26.1425C24.7352 26.6822 24.8578 27.1738 25.1038 27.6171C25.3494 28.0605 25.6793 28.414 26.0928 28.6773C26.5064 28.9411 26.965 29.0724 27.4685 29.0724C27.972 29.0724 28.4305 28.9411 28.8441 28.6773C29.2577 28.414 29.5871 28.0605 29.8331 27.6171C30.0787 27.1738 30.2018 26.6822 30.2018 26.1425V18.4708C30.2018 17.9387 30.6045 17.507 31.1009 17.507C31.5973 17.507 32 17.9387 32 18.4708V26.1425Z"
+        fill="#7A818B"
+      />
+      <path
+        d="M12 17.8575C12 16.9708 12.204 16.1582 12.6114 15.4191C13.0193 14.68 13.5645 14.0921 14.2478 13.6554C14.9311 13.2181 15.6925 13 16.5315 13C17.2749 13 17.9644 13.1735 18.5995 13.5204C19.235 13.8674 19.7626 14.3361 20.1819 14.9276C20.5898 14.3361 21.1113 13.8674 21.7464 13.5204C22.3819 13.1735 23.0714 13 23.8143 13C24.6415 13 25.3996 13.2181 26.0891 13.6554C26.7785 14.0921 27.327 14.68 27.7344 15.4191C28.1423 16.1577 28.3458 16.9708 28.3458 17.8575V25.5292C28.3458 26.0613 27.9431 26.493 27.4467 26.493C26.9503 26.493 26.5476 26.0613 26.5476 25.5292V17.8575C26.5476 17.3178 26.425 16.8262 26.179 16.3829C25.9334 15.9396 25.6035 15.586 25.1899 15.3227C24.7764 15.059 24.3178 14.9276 23.8143 14.9276C23.3108 14.9276 22.8736 15.0529 22.4656 15.3035C22.0582 15.554 21.7312 15.8914 21.4856 16.3154C21.24 16.7395 21.1051 17.2082 21.081 17.7226V25.5292C21.081 26.0613 20.6783 26.493 20.1819 26.493H20.1639C19.6675 26.493 19.2648 26.0613 19.2648 25.5292V17.8575C19.2648 17.3178 19.1422 16.8262 18.8962 16.3829C18.6506 15.9396 18.3207 15.586 17.9072 15.3227C17.4936 15.059 17.035 14.9276 16.5315 14.9276C16.028 14.9276 15.5695 15.059 15.1559 15.3227C14.7423 15.586 14.4129 15.9396 14.1669 16.3829C13.9213 16.8262 13.7982 17.3178 13.7982 17.8575V25.5292C13.7982 26.0613 13.3955 26.493 12.8991 26.493C12.4027 26.493 12 26.0613 12 25.5292V17.8575Z"
+        fill="#7A818B"
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_352_10">
+        <rect width="20" height="18" fill="white" transform="translate(12 13)" />
+      </clipPath>
+    </defs>
+  </svg>
+)
+
+export default MutationFallbackIcon
