+++ libs/shared-components/src/mini-overlay/assets/icons/StopCenter.tsx
@@ -1,0 +1,38 @@
+import React from 'react'
+
+const StopCenter = () => (
+  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <g clipPath="url(#clip0_194_487)">
+      <mask
+        id="path-1-outside-1_194_487"
+        maskUnits="userSpaceOnUse"
+        x="0"
+        y="0"
+        width="24"
+        height="24"
+        fill="black"
+      >
+        <rect fill="white" width="24" height="24" />
+        <path d="M12 2C10.6868 2 9.38642 2.25866 8.17317 2.7612C6.95991 3.26375 5.85752 4.00035 4.92893 4.92893C3.05357 6.8043 2 9.34784 2 12C2 14.6522 3.05357 17.1957 4.92893 19.0711C5.85752 19.9997 6.95991 20.7362 8.17317 21.2388C9.38642 21.7413 10.6868 22 12 22C14.6522 22 17.1957 20.9464 19.0711 19.0711C20.9464 17.1957 22 14.6522 22 12C22 10.6868 21.7413 9.38642 21.2388 8.17317C20.7362 6.95991 19.9997 5.85752 19.0711 4.92893C18.1425 4.00035 17.0401 3.26375 15.8268 2.7612C14.6136 2.25866 13.3132 2 12 2ZM9 9H15V15H9" />
+        <path d="M8 8H16V16H8V8Z" />
+      </mask>
+      <path
+        d="M12 2C10.6868 2 9.38642 2.25866 8.17317 2.7612C6.95991 3.26375 5.85752 4.00035 4.92893 4.92893C3.05357 6.8043 2 9.34784 2 12C2 14.6522 3.05357 17.1957 4.92893 19.0711C5.85752 19.9997 6.95991 20.7362 8.17317 21.2388C9.38642 21.7413 10.6868 22 12 22C14.6522 22 17.1957 20.9464 19.0711 19.0711C20.9464 17.1957 22 14.6522 22 12C22 10.6868 21.7413 9.38642 21.2388 8.17317C20.7362 6.95991 19.9997 5.85752 19.0711 4.92893C18.1425 4.00035 17.0401 3.26375 15.8268 2.7612C14.6136 2.25866 13.3132 2 12 2ZM9 9H15V15H9"
+        fill="#02193A"
+      />
+      <path d="M8 8H16V16H8V8Z" fill="white" />
+      <path
+        d="M2 12H0H2ZM12 22V24V22ZM15 9H17V7H15V9ZM15 15V17H17V15H15ZM8 8V6H6V8H8ZM16 8H18V6H16V8ZM16 16V18H18V16H16ZM8 16H6V18H8V16ZM12 0C10.4241 0 8.86371 0.310389 7.4078 0.913446L8.93853 4.60896C9.90914 4.20693 10.9494 4 12 4V0ZM7.4078 0.913446C5.95189 1.5165 4.62902 2.40042 3.51472 3.51472L6.34315 6.34315C7.08601 5.60028 7.96793 5.011 8.93853 4.60896L7.4078 0.913446ZM3.51472 3.51472C1.26428 5.76516 0 8.8174 0 12H4C4 9.87827 4.84285 7.84344 6.34315 6.34315L3.51472 3.51472ZM0 12C0 15.1826 1.26428 18.2348 3.51472 20.4853L6.34315 17.6569C4.84285 16.1566 4 14.1217 4 12H0ZM3.51472 20.4853C4.62902 21.5996 5.95189 22.4835 7.4078 23.0866L8.93853 19.391C7.96793 18.989 7.08602 18.3997 6.34315 17.6569L3.51472 20.4853ZM7.4078 23.0866C8.86371 23.6896 10.4241 24 12 24V20C10.9494 20 9.90914 19.7931 8.93853 19.391L7.4078 23.0866ZM12 24C15.1826 24 18.2348 22.7357 20.4853 20.4853L17.6569 17.6569C16.1566 19.1571 14.1217 20 12 20V24ZM20.4853 20.4853C22.7357 18.2348 24 15.1826 24 12H20C20 14.1217 19.1571 16.1566 17.6569 17.6569L20.4853 20.4853ZM24 12C24 10.4241 23.6896 8.86371 23.0866 7.4078L19.391 8.93853C19.7931 9.90914 20 10.9494 20 12H24ZM23.0866 7.4078C22.4835 5.95189 21.5996 4.62902 20.4853 3.51472L17.6569 6.34315C18.3997 7.08602 18.989 7.96793 19.391 8.93853L23.0866 7.4078ZM20.4853 3.51472C19.371 2.40042 18.0481 1.5165 16.5922 0.913446L15.0615 4.60896C16.0321 5.011 16.914 5.60028 17.6569 6.34315L20.4853 3.51472ZM16.5922 0.913446C15.1363 0.310389 13.5759 0 12 0V4C13.0506 4 14.0909 4.20693 15.0615 4.60896L16.5922 0.913446ZM9 11H15V7H9V11ZM13 9V15H17V9H13ZM15 13H9V17H15V13ZM8 10H16V6H8V10ZM14 8V16H18V8H14ZM16 14H8V18H16V14ZM10 16V8H6V16H10Z"
+        fill="white"
+        mask="url(#path-1-outside-1_194_487)"
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_194_487">
+        <rect width="24" height="24" fill="white" />
+      </clipPath>
+    </defs>
+  </svg>
+)
+
+export default StopCenter
