+++ libs/shared-components/src/mini-overlay/assets/icons/StopTop.tsx
@@ -1,0 +1,31 @@
+import React from 'react'
+
+const StopTop = () => (
+  <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <mask
+      id="path-1-outside-1_257_34"
+      maskUnits="userSpaceOnUse"
+      x="0.166687"
+      y="0.166672"
+      width="16"
+      height="16"
+      fill="black"
+    >
+      <rect fill="white" x="0.166687" y="0.166672" width="16" height="16" />
+      <path d="M8.00002 2.16667C7.23398 2.16667 6.47543 2.31756 5.7677 2.61071C5.05997 2.90386 4.41691 3.33354 3.87523 3.87522C2.78127 4.96918 2.16669 6.45291 2.16669 8.00001C2.16669 9.5471 2.78127 11.0308 3.87523 12.1248C4.41691 12.6665 5.05997 13.0962 5.7677 13.3893C6.47543 13.6825 7.23398 13.8333 8.00002 13.8333C9.54712 13.8333 11.0308 13.2188 12.1248 12.1248C13.2188 11.0308 13.8334 9.5471 13.8334 8.00001C13.8334 7.23396 13.6825 6.47542 13.3893 5.76769C13.0962 5.05995 12.6665 4.41689 12.1248 3.87522C11.5831 3.33354 10.9401 2.90386 10.2323 2.61071C9.52461 2.31756 8.76607 2.16667 8.00002 2.16667ZM6.25002 6.25001H9.75002V9.75001H6.25002" />
+      <path d="M5.66669 5.66667H10.3334V10.3333H5.66669V5.66667Z" />
+    </mask>
+    <path
+      d="M8.00002 2.16667C7.23398 2.16667 6.47543 2.31756 5.7677 2.61071C5.05997 2.90386 4.41691 3.33354 3.87523 3.87522C2.78127 4.96918 2.16669 6.45291 2.16669 8.00001C2.16669 9.5471 2.78127 11.0308 3.87523 12.1248C4.41691 12.6665 5.05997 13.0962 5.7677 13.3893C6.47543 13.6825 7.23398 13.8333 8.00002 13.8333C9.54712 13.8333 11.0308 13.2188 12.1248 12.1248C13.2188 11.0308 13.8334 9.5471 13.8334 8.00001C13.8334 7.23396 13.6825 6.47542 13.3893 5.76769C13.0962 5.05995 12.6665 4.41689 12.1248 3.87522C11.5831 3.33354 10.9401 2.90386 10.2323 2.61071C9.52461 2.31756 8.76607 2.16667 8.00002 2.16667ZM6.25002 6.25001H9.75002V9.75001H6.25002"
+      fill="#F43024"
+    />
+    <path d="M5.66669 5.66667H10.3334V10.3333H5.66669V5.66667Z" fill="white" />
+    <path
+      d="M2.16669 8.00001H0.166687H2.16669ZM8.00002 13.8333V15.8333V13.8333ZM9.75002 6.25001H11.75V4.25001H9.75002V6.25001ZM9.75002 9.75001V11.75H11.75V9.75001H9.75002ZM5.66669 5.66667V3.66667H3.66669V5.66667H5.66669ZM10.3334 5.66667H12.3334V3.66667H10.3334V5.66667ZM10.3334 10.3333V12.3333H12.3334V10.3333H10.3334ZM5.66669 10.3333H3.66669V12.3333H5.66669V10.3333ZM8.00002 0.166672C6.97133 0.166672 5.95272 0.369287 5.00233 0.762949L6.53307 4.45847C6.99815 4.26582 7.49662 4.16667 8.00002 4.16667L8.00002 0.166672ZM5.00233 0.762949C4.05195 1.15661 3.18841 1.73361 2.46102 2.461L5.28944 5.28943C5.6454 4.93347 6.06799 4.65111 6.53307 4.45847L5.00233 0.762949ZM2.46102 2.461C0.991982 3.93004 0.166687 5.92248 0.166687 8.00001L4.16669 8.00001C4.16669 6.98334 4.57055 6.00832 5.28944 5.28943L2.46102 2.461ZM0.166687 8.00001C0.166687 10.0775 0.991982 12.07 2.46102 13.539L5.28944 10.7106C4.57055 9.99169 4.16669 9.01667 4.16669 8.00001L0.166687 8.00001ZM2.46102 13.539C3.18841 14.2664 4.05195 14.8434 5.00233 15.2371L6.53307 11.5415C6.06799 11.3489 5.6454 11.0665 5.28944 10.7106L2.46102 13.539ZM5.00233 15.2371C5.95272 15.6307 6.97133 15.8333 8.00002 15.8333L8.00002 11.8333C7.49662 11.8333 6.99815 11.7342 6.53307 11.5415L5.00233 15.2371ZM8.00002 15.8333C10.0775 15.8333 12.07 15.008 13.539 13.539L10.7106 10.7106C9.99171 11.4295 9.01668 11.8333 8.00002 11.8333L8.00002 15.8333ZM13.539 13.539C15.0081 12.07 15.8334 10.0775 15.8334 8.00001H11.8334C11.8334 9.01667 11.4295 9.99169 10.7106 10.7106L13.539 13.539ZM15.8334 8.00001C15.8334 6.97132 15.6307 5.9527 15.2371 5.00232L11.5416 6.53305C11.7342 6.99813 11.8334 7.4966 11.8334 8.00001H15.8334ZM15.2371 5.00232C14.8434 4.05193 14.2664 3.18839 13.539 2.461L10.7106 5.28943C11.0666 5.64539 11.3489 6.06797 11.5416 6.53305L15.2371 5.00232ZM13.539 2.461C12.8116 1.73361 11.9481 1.15661 10.9977 0.762949L9.46697 4.45847C9.93206 4.65111 10.3546 4.93347 10.7106 5.28943L13.539 2.461ZM10.9977 0.762949C10.0473 0.369287 9.02871 0.166672 8.00002 0.166672L8.00002 4.16667C8.50342 4.16667 9.00189 4.26582 9.46697 4.45847L10.9977 0.762949ZM6.25002 8.25001H9.75002V4.25001H6.25002V8.25001ZM7.75002 6.25001V9.75001H11.75V6.25001H7.75002ZM9.75002 7.75001H6.25002V11.75H9.75002V7.75001ZM5.66669 7.66667H10.3334V3.66667H5.66669V7.66667ZM8.33335 5.66667V10.3333H12.3334V5.66667H8.33335ZM10.3334 8.33334H5.66669V12.3333H10.3334V8.33334ZM7.66669 10.3333V5.66667H3.66669V10.3333H7.66669Z"
+      fill="white"
+      mask="url(#path-1-outside-1_257_34)"
+    />
+  </svg>
+)
+
+export default StopTop
