+++ libs/shared-components/src/mini-overlay/assets/icons/index.ts
@@ -1,0 +1,8 @@
+import MutationFallbackIcon from './MutationFallbackIcon'
+import ArrowIcon from './Arrow'
+import StopTopIcon from './StopTop'
+import PlayCenterIcon from './PlayCenter'
+import StopCenterIcon from './StopCenter'
+import BellIcon from './Bell'
+
+export { MutationFallbackIcon, ArrowIcon, StopTopIcon, PlayCenterIcon, StopCenterIcon, BellIcon }
