+++ libs/shared-components/src/notifications/assets/icons/GreenBadge.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const GreenBadge = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="14" height="15" viewBox="0 0 14 15" fill="none">
+    <rect y="0.5" width="14" height="14" rx="7" fill="#19CEAE" />
+    <path
+      d="M4.83492 11.4998C4.52013 11.5003 4.21534 11.3893 3.97452 11.1865C3.73371 10.9837 3.57241 10.7023 3.51919 10.3919C3.46597 10.0816 3.52427 9.76244 3.68377 9.49098C3.84327 9.21953 4.09366 9.0133 4.39061 8.90881V6.09119C4.09403 5.98635 3.84406 5.78001 3.68488 5.50864C3.5257 5.23727 3.46756 4.91835 3.52073 4.60825C3.57391 4.29815 3.73497 4.01684 3.97546 3.81403C4.21594 3.61123 4.52037 3.5 4.83492 3.5C5.14947 3.5 5.45389 3.61123 5.69438 3.81403C5.93486 4.01684 6.09593 4.29815 6.1491 4.60825C6.20227 4.91835 6.14413 5.23727 5.98495 5.50864C5.82577 5.78001 5.5758 5.98635 5.27923 6.09119V6.16674C5.27923 6.40247 5.37285 6.62855 5.5395 6.79524C5.70614 6.96193 5.93217 7.05558 6.16784 7.05558H8.8337C9.30505 7.05558 9.7571 7.24287 10.0904 7.57625C10.4237 7.90963 10.6109 8.36179 10.6109 8.83326V8.90881C10.9075 9.01365 11.1575 9.21999 11.3167 9.49136C11.4758 9.76273 11.534 10.0816 11.4808 10.3917C11.4276 10.7018 11.2666 10.9832 11.0261 11.186C10.7856 11.3888 10.4812 11.5 10.1666 11.5C9.85208 11.5 9.54765 11.3888 9.30717 11.186C9.06668 10.9832 8.90562 10.7018 8.85244 10.3917C8.79927 10.0816 8.85741 9.76273 9.01659 9.49136C9.17577 9.21999 9.42574 9.01365 9.72232 8.90881V8.83326C9.72232 8.59753 9.62869 8.37145 9.46205 8.20476C9.2954 8.03807 9.06937 7.94442 8.8337 7.94442H6.16784C5.85591 7.94454 5.54944 7.86254 5.27923 7.70666V8.90926C5.57533 9.01443 5.82479 9.22082 5.9836 9.49202C6.14241 9.76323 6.20036 10.0818 6.14723 10.3916C6.0941 10.7014 5.9333 10.9824 5.6932 11.1852C5.45311 11.3879 5.14914 11.4993 4.83492 11.4998Z"
+      fill="white"
+    />
+  </svg>
+)
+
+export default GreenBadge
