+++ libs/shared-components/src/notifications/assets/icons/NotificationClose.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+
+const NotificationClose = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 12 12" fill="none">
+    <path
+      d="M6.75 9.5H2.5C2.23478 9.5 1.98043 9.39464 1.79289 9.20711C1.60536 9.01957 1.5 8.76522 1.5 8.5V3.5C1.5 3.23478 1.60536 2.98043 1.79289 2.79289C1.98043 2.60536 2.23478 2.5 2.5 2.5H9.5C9.76522 2.5 10.0196 2.60536 10.2071 2.79289C10.3946 2.98043 10.5 3.23478 10.5 3.5V6.5"
+      stroke="#7A818B"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M1.5 3.5L6 6.5L10.5 3.5M11 11L8.5 8.5M8.5 11L11 8.5"
+      stroke="#7A818B"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+)
+
+export default NotificationClose
