+++ libs/shared-components/src/notifications/assets/icons/RedBadge.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+
+const RedBadge = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="14" height="15" viewBox="0 0 14 15" fill="none">
+    <g clipPath="url(#clip0_452_830)">
+      <path
+        d="M7 0.5C8.85652 0.5 10.637 1.2375 11.9497 2.55025C13.2625 3.86301 14 5.64348 14 7.5C14 9.35652 13.2625 11.137 11.9497 12.4497C10.637 13.7625 8.85652 14.5 7 14.5C5.14348 14.5 3.36301 13.7625 2.05025 12.4497C0.737498 11.137 0 9.35652 0 7.5C0 5.64348 0.737498 3.86301 2.05025 2.55025C3.36301 1.2375 5.14348 0.5 7 0.5ZM4.8125 11.2188C5.1808 11.2194 5.53701 11.0873 5.81582 10.8467C6.09463 10.606 6.27736 10.273 6.33052 9.90851C6.38368 9.54406 6.3037 9.17267 6.10524 8.8624C5.90679 8.55214 5.60315 8.3238 5.25 8.21925V6.562C5.60079 6.45519 5.90167 6.22618 6.09806 5.91652C6.29445 5.60686 6.37334 5.23706 6.32042 4.87421C6.2675 4.51137 6.08626 4.17951 5.8096 3.93885C5.53294 3.69818 5.17918 3.56466 4.8125 3.5625C4.4442 3.56183 4.08799 3.69392 3.80918 3.93457C3.53037 4.17522 3.34764 4.5083 3.29448 4.87274C3.24132 5.23719 3.3213 5.60858 3.51976 5.91885C3.71822 6.22911 4.02185 6.45744 4.375 6.562V8.21925C4.02421 8.32606 3.72333 8.55507 3.52694 8.86473C3.33055 9.17439 3.25166 9.54419 3.30458 9.90703C3.3575 10.2699 3.53874 10.6017 3.8154 10.8424C4.09206 11.0831 4.44582 11.2166 4.8125 11.2188ZM9.1875 6.84375C9.07147 6.84375 8.96019 6.88984 8.87814 6.97189C8.79609 7.05394 8.75 7.16522 8.75 7.28125V8.21925C8.39921 8.32606 8.09833 8.55507 7.90194 8.86473C7.70555 9.17439 7.62666 9.54419 7.67958 9.90703C7.7325 10.2699 7.91374 10.6017 8.1904 10.8424C8.46706 11.0831 8.82082 11.2166 9.1875 11.2188C9.5558 11.2194 9.91201 11.0873 10.1908 10.8467C10.4696 10.606 10.6524 10.273 10.7055 9.90851C10.7587 9.54406 10.6787 9.17267 10.4802 8.8624C10.2818 8.55214 9.97815 8.3238 9.625 8.21925V7.28125C9.625 7.16522 9.57891 7.05394 9.49686 6.97189C9.41481 6.88984 9.30353 6.84375 9.1875 6.84375ZM10.2856 4.42613C10.3653 4.34361 10.4094 4.2331 10.4084 4.11839C10.4074 4.00368 10.3614 3.89395 10.2803 3.81283C10.1992 3.73171 10.0894 3.6857 9.97474 3.68471C9.86003 3.68371 9.74951 3.72781 9.667 3.8075L9.20325 4.27212L8.73863 3.8075C8.65611 3.72781 8.5456 3.68371 8.43089 3.68471C8.31618 3.6857 8.20645 3.73171 8.12533 3.81283C8.04421 3.89395 7.9982 4.00368 7.9972 4.11839C7.99621 4.2331 8.04031 4.34361 8.12 4.42613L8.58463 4.89075L8.12 5.3545C8.04031 5.43701 7.99621 5.54753 7.9972 5.66224C7.9982 5.77695 8.04421 5.88668 8.12533 5.9678C8.20645 6.04891 8.31618 6.09492 8.43089 6.09592C8.5456 6.09692 8.65611 6.05282 8.73863 5.97312L9.20325 5.50938L9.667 5.97312C9.74951 6.05282 9.86003 6.09692 9.97474 6.09592C10.0894 6.09492 10.1992 6.04891 10.2803 5.9678C10.3614 5.88668 10.4074 5.77695 10.4084 5.66224C10.4094 5.54753 10.3653 5.43701 10.2856 5.3545L9.82188 4.89075L10.2856 4.42613Z"
+        fill="#DB504A"
+      />
+    </g>
+    <defs>
+      <clipPath id="clip0_452_830">
+        <rect width="14" height="14" fill="white" transform="translate(0 0.5)" />
+      </clipPath>
+    </defs>
+  </svg>
+)
+
+export default RedBadge
