+++ libs/shared-components/src/notifications/assets/icons/index.ts
@@ -1,0 +1,7 @@
+import BlueBadge from "./BlueBadge";
+import RedBadge from "./RedBadge";
+import GreenBadge from "./GreenBadge";
+import Collapse from "./Collapse";
+import NotificationMessage from "./NotificationMessage";
+import NotificationClose from "./NotificationClose";
+export { BlueBadge, RedBadge, GreenBadge, Collapse, NotificationMessage, NotificationClose }
